Switch to node-cassandra-cql driver, fix HTTP status codes
Switch to the binary node-cassandra-sql driver for better performance and
better map/set/list support. This roughly doubles the throughput on my laptop
to about 2GBit on my laptop when testing with the Obama wikitext. A real
server should be able to fill a 10GBit connection. Small requests remain
unchanged around 1800req/s.