Test different types, and add proper coercion methods
So far we didn't actually test all supported data types. Adding a new set of
tests exposed several issues.

This patch implements proper Cassandra VarInt encoding and decoding (which is
sadly not handled in the cassandra driver). For now it uses native JS Numbers,
which is a limitation as bitwise arithmetic is limited to 32 bits. It makes
sense to switch the varint encoding to the bignum package in a follow-up
patch.

Another issue with floats was exposed in the cassandra driver. This is tracked
as https://datastax-oss.atlassian.net/browse/NODEJS-50 upstream. Until this is
fixed, float support has been disabled.

Decimal support is not yet implemented properly, so has also been disabled.