More aggressive reconnection in the cassandra driver
In load testing I observed at least one worker with connections to only two
out of three cassandra nodes. This worker was clearly not making progress in
its connection with cassandra, and as a result was building up memory despite
only having around 90 client connections associated with it at any time. It
seemed that client connections would time out, but the state associated with
the connection would still remain in memory until the cassandra connection
would clear (which it never did). Other workers on the same node had
connections to all three cassandra nodes, and were making good progress with
little memory used.

This patch attempts to help the situation a bit by at least trying to
reconnect more aggressively. That might have helped that client to
re-establish a connection to the third cassandra node more quickly. I suspect
though that stuck TCP connections are not detected in a speedy manner inside
the cassandra driver. There might be a bug somewhere in there.