Added a script to send requests directly to the RB storage
RESTBase depends on its storage component (restbase-cassandra) and uses
it internally. However, there is no way for the user to interact with it
directly. This script fills the gap by using the same mechanisms as the
test scripts. Concretely, it constructs a router able to translate URLs
(as would be gotten from RESTBase) to storage back-end method calls. The
user must provide the path to route the request to, and may additionally
provide the request method and body.