Added better input data parsing and evaluation
Do not blindly assume the JSON is well-formed. Instead, catch any errors
during the parse and try to eval() the input data. While this seems like
a risky move to make, it gives us greater flexibility on the command
line. Ultimately, the user is responsible for the code they are
executing, and therefore it makes sense to 'let them do what they want'.
Furthermore, direct-request-js is more of a dev script and should not be
used in production unless really needed.