Enable TCP keepalive in the cassandra driver
Temporarily use a patched version of the cassandra driver with TCP keepalive
enabled. We are seeing cassandra connection hangs in production, in particular
after an unclean cassandra shutdown (OOM). This then causes client connections
to pile up in the worker, which in turn uses a lot of memory until the client
runs out of memory as well & is restarted by the coordinator.

Lets see if this helps.