T84900: Use a JS Object for storing conversion functions
Until now, conversion functions were stored in an array of JS Objects,
which is sub-optimal since in 99% of the cases where a conversion is
needed, the attribute's name is known in advance. The new outline allows
one to invoke the conversion function directly by using
schema.conversions.<attName>.(write|read)().