T84900: Minor: Refactor conversion info a bit
Until now, when detecting if a conversion method is needed for a
schema's attribute, a new object would be created each time. Now, we
define a static dbu.conversions object which contains all of the
conversion definitions for known types, and then reference each of its
keys when determining an attribute's conversion rules. As an added
bonus, this will allow us to easily define conversion schemes for
set-typed attributes.