Test router: turn module exported object into a routeswitch handler
Instead of exporting handlers directly, the module now exports a spec
and an operation object. However, in the tests we use routeswitch which
expects concrete handler methods. This commit translates the module's
exported object into a handler object, checking all of the needed
operationId's have been exported by the module in the process.