Move transforms into get()
With the implicit decoding support in v2 of the cassandra driver we can focus
our transforms on the mapping of Cassandra types to JSON. This is only needed
on the way out, so this patch moves the result transforms to be only applied
in `get()`. All internal users use the private `_get()` worker method, and
thus get native Cassandra types. This is more efficient and convenient
especially when processing secondary index updates.

Also:
- Remove the cassandra-codec dependency. This means that we now have no binary
  module dependencies.
- Clarify and reformat table creation tests