Drop settle time with v2 driver & log table creation
In local testing against a single node the v2 driver reliably created
thousands of tables without any settle time, so it seems that the fixes in
schema update handling are effective. This patch just sets the delay to zero
(so that we can easily bump it back up in case it's still needed with multiple
nodes). After successful testing we can remove the delay statements
completely.

Also log table creations as a warning. These are important events, so should
be in the logs in the default 'warn' log level.