Increase the default prepared statement limit to 50k
There seems to be no way to parametrize keyspace or column family (table)
names, which means that we end up with one set of prepared statements per
domain. At 800 domains that far exceeds the default maxPrepared limit of 500,
so bump it to 50000. This value will still easily catch a lack of
parametrization on values, and should (hopefully) still result in reasonable
resource consumption on server & client.