Stop creating native Cassandra indexes on _domain
A bug in Cassandra 2.1 causes inserts in tables with native secondary indexes
on _domain to fail if the size of any column value exceeds a fairly low
threshold (16k IIRC). This is a problem especially on tables backing content
tables like those behind key_rev_value storage.

This patch stops creating those secondary indexes by default. This will break
the hacky revision and title listings in RESTBase, which we should disable or
replace with API requests.

Next steps in the backend:

- Automatically drop existing secondary indexes on `_domain`. Naming scheme is
  `<table_name>_<column_name>_idx_<counter>`, with the main difficulty being
  the counter value.