use json-stable-stringify for schema hashing
json-stable-stringify does what we want here, and simplifies things quite a
bit.  There is one problem however, a bug[1] that causes false positive
circular reference exceptions.  A fix exists that Works For Me, so as a
temporary solution, this change uses a third-party branch to satisfy the
dependency.  If this is acceptable as a stop-gap, I'll open an issue to
keep track of this, and see to it that it's not forgotten.

[1]: https://github.com/substack/json-stable-stringify/issues/14
[2]: https://github.com/substack/json-stable-stringify/pull/15

Bug: T96612