back-end versioning
Given the case where *both* a backend version migration, and a table
storage schema migration are needed, it stands to reason that the latter
might depend on the former.  Because of this, the two cases need to be
treated separately, and table schema migration must occur after any backend
migration has completed.

This changeset introduces a `_backend_version` attribute to the schema blob
but does not consider it when generating the hash.  DB#_createTable
first evaluates `_backend_version`, performs a back-end migration if
necessary, then compares hashes and performs a table schema migration if
necessary.  Finally, if any changes were made, the persisted schema blob
is updated.

Bug: T97845