Bug fix: Normalise asynchronous background updates
When a record is added/updated, the asynchronous updates kick in to
update the secondary indices and enforce the declared retention policy.
However, the original request's attributes are JS types, while all of
the retrieved DB rows for which these actions are applied are in native,
Cassandra representation, while the code supposed both were JS.

This commit fixes this issue by adding an additional parameter to
`buildPutQuery()` and `buildCondition()` allowing users to supply their own
conversion function which is applied to a request's query's attributes.
This feature is used by the revision policy and the secondary index
rebuilder, which supply `dbu.noConvert()`, a dummy conversion function
simply returning the value passed in. If this extra parameter is not
passed to `buildPutQuery()`, the default conversion scheme is used, thus
not affecting the workflow of the other parts of the system.

Additionally, the request query attributes passed to the constructors of
`RevisionPolicyManager` and `IndexRebuilder` are converted to the native
Cassandra value types to avoid mixing Cassandra and JS types.