Minor fix: Check query.if more strictly
Now, if query.if is supplied and is a string, it must be "not exists".
If it's not a string, then it's required to be an object. In both
instances if the condition has not been met, an error is thrown. This
commit also adds two more tests to the test suite to enforce such
behaviour.