Make background updates synchronous
With idempotent updates it is better to retry the put at a higher level if
necessary. Consumers can still choose to ignore failures, but now have a
chance to repeat requests in case any part of the request processing fails.

This change should provide better back-pressure in write overload situations
by slowing those requests down until the request is fully processed. As a
consequence, it should reduce the memory build-up in restbase when writes are
slow.

Additionaly, the patch also cleans up the condition for when background
updates are run.