Do not apply retention policies to newer and current data
Applying the retention policy to newer / current rows is badly broken, as we
can end up deleting the current row whenever a newer row is found. This patch
applies the retention policy only to older rows, and compensates for not
processing the current row by starting the row count with 1.