Un-do the _fetchSchema retry logic
It turns out that it actually makes more sense to return the request without a
schema if a "Keyspace <x> does not exist" error was returned by the original
query. It's quite clear that system.schema_columnfamilies and the driver's
idea of the schema often disagree, and it seems to be more robust to trust the
driver more. By returning a null schema to a createTable request, the table is
re-created, which fixes up any inconsistencies.