Update parameters for DTCS tables & improve backend migration
This patch updates DTCS parameters for current stable Cassandra (2.2.4 and
2.1.12) by setting max_window_size_seconds explicitly, and dropping some less
important parameters. The exact settings aren't completely final yet, so we
might want to wait with applying this for a week or two until we have gotten a
bit more data on good defaults. These settings in particular are somewhat
questionable:

- unchecked_tombstone_compaction is disabled on some high-traffic prod tables,
  but we might want to revisit this in light of 2.1.12.
- tombstone_threshold: Set to 0.15 on high-traffic prod tables, but might also
  want to lower this in light of the increased compaction headroom in
  2.1.12. The current patch specified 0.1, which might be a good balance.
- max_window_size_seconds is set to 60 days in large production tables & this
  patch. We might want to increase this.

This patch also provides a backend migration pass to apply these parameter
changes. To make it easy to call other schema migrators, the main
BackendMigrator code now has all the logic needed.