GET query caching
A significant portion of our CPU time was spent in the repeated construction
of GET queries. This patch adds a cache for GET queries, avoiding the need to
repeatedly construct CQL from scratch.

In the process, I also found a small bug in _deleteOlderContent. The logic
there dependend on the order of Object.keys(), which is rather fragile.